wasi: share extreme-mtime fixture and dedupe test runners
Address review feedback:
- Implement run() via run_with_workspace_setup() in p1, p2-sync, and
  p2-async so the runners are not duplicated.
- Move host extreme-mtime workspace seeding into
  store::prepare_extreme_mtime_fixture for shared use.
- Drop the unit test module in filesystem.rs now that integration
  coverage exists.

cargo test -p wasmtime-wasi --features p1,p2 --test all p1_stat_extreme_host_mtime
