+++ crates/wasi/src/filesystem.rs
@@ -1171,54 +1171,3 @@ impl WasiFilesystemCtxView<'_> {
         Ok(results)
     }
 }
-
-#[cfg(test)]
-mod test {
-    use super::*;
-    use crate::clocks::Datetime;
-    use std::io::{Read, Write};
-    use std::time::{Duration, SystemTime};
-
-    #[test]
-    fn datetime_try_from_extreme_system_time() {
-        let extreme = SystemTime::UNIX_EPOCH
-            .checked_sub(Duration::from_secs((i64::MAX as u64) + 1))
-            .unwrap();
-        assert!(Datetime::try_from(extreme).ok().is_none());
-    }
-
-    #[test]
-    fn host_prepared_extreme_mtime_stats_and_reads() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("extreme.dat");
-        std::fs::File::create(&path)
-            .unwrap()
-            .write_all(b"hello")
-            .unwrap();
-
-        let extreme = SystemTime::UNIX_EPOCH
-            .checked_sub(Duration::from_secs((i64::MAX as u64) + 1))
-            .unwrap();
-        let f = std::fs::OpenOptions::new().write(true).open(&path).unwrap();
-        let times = std::fs::FileTimes::new()
-            .set_modified(extreme)
-            .set_accessed(extreme);
-        // Platforms may reject or clamp extreme times.
-        let _ = f.set_times(times);
-        drop(f);
-
-        let f = std::fs::File::open(&path).unwrap();
-        let stat = sys::stat(&f).expect("stat");
-        let mut buf = Vec::new();
-        std::fs::File::open(&path)
-            .unwrap()
-            .read_to_end(&mut buf)
-            .unwrap();
-        assert_eq!(buf, b"hello");
-        let _ = (
-            stat.data_access_timestamp,
-            stat.data_modification_timestamp,
-            stat.status_change_timestamp,
-        );
-    }
-}
