+++ crates/wasi/tests/all/p2/async_.rs
@@ -8,23 +8,7 @@ use wasmtime_wasi::p2::add_to_linker_async;
 use wasmtime_wasi::p2::bindings::Command;
 
 async fn run(path: &str, with_builder: impl FnOnce(&mut WasiCtxBuilder)) -> Result<()> {
-    let path = Path::new(path);
-    let name = path.file_stem().unwrap().to_str().unwrap();
-    let engine = test_programs_artifacts::engine(|_config| {});
-    let mut linker = Linker::new(&engine);
-    add_to_linker_async(&mut linker)?;
-
-    let (mut store, _td) = Ctx::new(&engine, name, |builder| {
-        with_builder(builder);
-        MyWasiCtx::new(builder.build())
-    })?;
-    let component = Component::from_file(&engine, path)?;
-    let command = Command::instantiate_async(&mut store, &component, &linker).await?;
-    command
-        .wasi_cli_run()
-        .call_run(&mut store)
-        .await?
-        .map_err(|()| wasmtime::format_err!("run returned a failure"))
+    run_with_workspace_setup(path, |_| Ok(()), with_builder).await
 }
 
 async fn run_with_workspace_setup(
@@ -98,23 +82,9 @@ async fn p1_fd_filestat_set() {
 }
 #[test_log::test(tokio::test(flavor = "multi_thread"))]
 async fn p1_stat_extreme_host_mtime() {
-    use std::fs::{File, FileTimes};
-    use std::io::Write;
-    use std::time::{Duration, SystemTime};
-
     run_with_workspace_setup(
         P1_STAT_EXTREME_HOST_MTIME_COMPONENT,
-        |dir| {
-            let path = dir.join("extreme.dat");
-            File::create(&path)?.write_all(b"hello")?;
-            let extreme = SystemTime::UNIX_EPOCH
-                .checked_sub(Duration::from_secs((i64::MAX as u64) + 1))
-                .expect("construct extreme SystemTime");
-            let f = File::options().write(true).open(&path)?;
-            let times = FileTimes::new().set_modified(extreme).set_accessed(extreme);
-            let _ = f.set_times(times);
-            Ok(())
-        },
+        crate::store::prepare_extreme_mtime_fixture,
         |_| {},
     )
     .await
